// src/api/database/fragments/types/Test.ts
/* tslint:disable */
/* eslint-disable */
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL fragment: Test
// ====================================================

export interface Test {
  __typename: "Test";
  id: any;
  index: number;
  input: string;
  output: string;
  isPublic: boolean | null;
  problemId: any;
  createdAt: any;
  updatedAt: any;
}